+++ test/test_bible_routes/test_language_routes.py
@@ -9,11 +9,11 @@ class TestLanguageAndScriptEndpoints:
         headers = {"Authorization": f"Bearer {regular_token1}"}
         response = client.get("/v3/language", headers=headers)
         assert response.status_code == 200
-        assert response.json() == [{"iso639": "eng", "name": "English"}]
+        assert response.json() == [{"iso639": "eng", "name": "english"}]
 
-    def test_list_scripts(self, client, regular_token1, setup_iso_data):
+    def test_list_scripts(self, client, regular_token1, test_db_session):
         # Test listing scripts
         headers = {"Authorization": f"Bearer {regular_token1}"}
         response = client.get("/v3/script", headers=headers)
         assert response.status_code == 200
-        assert response.json() == [{"iso15924": "Latn", "name": "Latin"}]
+        assert response.json() == [{"iso15924": "Latn", "name": "latin"}]
